requirements.txt
python-dotenv
requests
numpy
pyarrow
pandas
wxpython